Handle pipes being full
A process's output can fill its pipe.
While the process is running, we check if its pipe is full every PROCESSWAIT ms (default 100ms). If its full we empty the buffer by reading all of the available data.

Also add a PIPESZ constant so we know what the buffer size is.