Fix jobs output
Previously we would wait for jobs to finish one by one which means that:
1. One job could hold back the output of multiple jobs.
2. The output of jobs was only passed to CollectOutput after the job was done. This hides the true order that things occurred in.

Now we act much like the upstream Job_CatchChildren. We check if a job is done, if it is, we call JobFinish. Otherwise, we just call CollectOutput. This causes a busy wait while jobs are running, but without getting overlapped I/O involved, this is the best we can do.

Also remove the carriage return check in CollectOutput, it was added before a758fa7